+++ examples/hosts/app-integration/schema-upgrade/src/migrationTools/index.ts
@@ -3,5 +3,5 @@
  * Licensed under the MIT License.
  */
 
-export * from "./containerKillBit";
+export * from "./migrationTools";
 export * from "./interfaces";
