+++ examples/hosts/app-integration/schema-upgrade/src/migrationTools/interfaces.ts
